[build] Bump error prone from 2.9.0 to 2.32.0

+++ smack-integration-test/src/main/java/org/jivesoftware/smackx/httpfileupload/HttpFileUploadIntegrationTest.java
@@ -82,7 +82,7 @@ public class HttpFileUploadIntegrationTest extends AbstractSmackIntegrationTest 
         URL getUrl = hfumOne.uploadFile(file, new UploadProgressListener() {
             @Override
             public void onUploadProgress(long uploadedBytes, long totalBytes) {
-                double progress = uploadedBytes / totalBytes;
+                double progress = uploadedBytes / ((double) totalBytes);
                 LOGGER.fine("HTTP File Upload progress " + progress + "% (" + uploadedBytes + '/' + totalBytes + ')');
             }
         });
